+++ tests/stick_test_data.py
@@ -1344,6 +1344,19 @@ RESPONSE_MESSAGES = {
         + b"00000444"
         + b"00044020",
     ),
+    b"\x05\x05\x03\x030009333333333333333302047290\r\n": (
+        "Reset node request for 3333333333333333",
+        b"000000C1",  # Success ack
+        b"0000" + b"00F2" + b"3333333333333333",  # msg_id, reset_ack, mac
+    ),
+    b"\x05\x05\x03\x03001C009876543210123433333333333333331E2D\r\n": (
+        "Remove node request for 3333333333333333",
+        b"000000C1",  # Success ack
+        b"001D"  # msg_id
+        + b"0098765432101234"  # Circle + mac
+        + b"3333333333333333"
+        + b"01",  # status
+    ),
 }