+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Ongoing
 
+- PR [337](https://github.com/plugwise/python-plugwise-usb/pull/337): Improve node removal, remove and reset the node as executed by Source, and remove the cache-file.
 - PR [342](https://github.com/plugwise/python-plugwise-usb/pull/342): Improve node_type chaching.
 
 ## 0.46.0 - 2025-09-12